+++ reactos/ntoskrnl/io/iomgr/util.c
@@ -186,7 +186,7 @@ IoCheckEaBufferValidity(IN PFILE_FULL_EA_INFORMATION EaBuffer,
             EaBufferEnd->EaNameLength + EaBufferEnd->EaValueLength +
             FIELD_OFFSET(FILE_FULL_EA_INFORMATION, EaName[0]) + 1;
 
-        if (IntEaLength >= NextEaBufferOffset)
+        if ((ULONG)IntEaLength >= NextEaBufferOffset)
         {
             /* is the EaBufferName terminated with zero? */
             if (EaBufferEnd->EaName[EaBufferEnd->EaNameLength]==0)
